ocrec/requirements.txt
scikit_image==0.14.1
opencv_python==3.4.4.19
tqdm==4.28.1
torch==1.7.1
scipy==1.2.2
numpy==1.20.3
torchvision==0.8.2
Pillow==8.1.1
tensorboardX==1.5
editdistance==0.5.3
torchaudio==0.7.2
PyQt5==5.12.3
PyAutoGUI==0.9.52